+++ scripts/daemon/start.js
@@ -1,11 +1,11 @@
-var run = require('child_process').fork;
+var run = require('child_process').spawn;
 var gn = require('../../');
 var logger = gn.log.create('daemon-start');
 
 module.exports = function (path) {
 	gn.on('uncaughtException', gn.exit);
 	logger.info('starting monitor process...');
-	run(gn._root + 'scripts/daemon/monitor', ['start', path]);
+	var child = run(process.execPath, [gn._root + 'scripts/daemon/monitor', 'start', path], { detached: true, stdio: 'ignore' });
 	logger.info('monitor started...');
 	gn.exit();
 };