+++ core/process.js
@@ -2,7 +2,7 @@
 var EventEmitter = require('events').EventEmitter;
 var util = require('util');
 var cluster = require('cluster');
-var daemon;
+var daemon = require('./daemon');
 
 module.exports = Process;
 
@@ -46,8 +46,8 @@ Process.prototype.setup = function () {
 	this.log.info('running the process in none-cluster mode (pid: ' + process.pid + ')');
 
 	// set up daemon in case we need it
-	daemon = require('./daemon');
 	daemon.processObj = this;
+	daemon.setupDaemonService();
 	
 	this.emit('nocluster.setup');
 };
@@ -56,8 +56,8 @@ Process.prototype.setup = function () {
 Process.prototype.startClusterMode = function () {
 	if (cluster.isMaster) {
 		// set up daemon in case we need it
-		daemon = require('./daemon');
 		daemon.processObj = this;
+		daemon.setupDaemonService();
 		// set up master process
 		return this.setupMaster();
 	}